requirements.txt
anyascii==0.3.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
boto3==1.22.10
botocore==1.25.10
certifi==2021.10.8
charset-normalizer==2.0.11
dj-database-url==0.5.0
Django==4.0.2
django-cogwheels==0.3
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-filter==21.1
django-modelcluster==5.2
django-slick-reporting==0.6.3
django-storages==1.12.3
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jmespath==1.0.0
l18n==2021.3
openpyxl==3.0.9
Pillow==9.0.1
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
telepath==0.2
tzdata==2021.5
urllib3==1.26.8
wagtail==2.16
wagtailmenus==3.0.2
webencodings==0.5.1
whitenoise==6.0.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
